refactoring: 공통으로 사용되는 버튼 컴포넌트 파일 위치 변경
CancelButton, ConfirmButton 컴포넌트를 component/CommonButton/... 으로 이동
